libreactor: replace setup.c with helpers.c for the libreactor-server implementation
