+++ frameworks/C/libreactor/src-server/helpers.h
@@ -1,0 +1,6 @@
+#ifndef HELPERS_H_INCLUDED
+#define HELPERS_H_INCLUDED
+
+void setup();
+
+#endif /* HELPERS_H_INCLUDED */
