+++ frameworks/C/libreactor/src-server/main.c
@@ -1,15 +1,15 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include <stdint.h>
+#include <stdlib.h>
 #include <unistd.h>
-#include <string.h>
 #include <err.h>
 
 #include <dynamic.h>
 #include <reactor.h>
 #include <clo.h>
 
-#include "setup.h"
+#include "helpers.h"
+
 
 static core_status server_handler(core_event *event)
 {
